fix: extract LND errors to LndMobileErrors, refactor waitForLndReady fix typos
