+++ stores/InvoicesStore.ts
@@ -205,61 +205,50 @@ export default class InvoicesStore {
             unified: true,
             customPreimage,
             noLsp
-        })
-            .then(
-                (
-                    result:
-                        | { rHash: string; paymentRequest: string }
-                        | undefined
-                ) => {
-                    if (!result?.rHash || !result?.paymentRequest) {
-                        runInAction(() => {
-                            this.loading = false;
-                            this.creatingInvoice = false;
-                        });
-                        return;
-                    }
-                    const { rHash, paymentRequest } = result;
-                    if (
-                        BackendUtils.supportsOnchainReceiving() &&
-                        !skipOnchain
-                    ) {
-                        return this.getNewAddress(
-                            addressType
-                                ? { type: addressType, unified: true }
-                                : { unified: true }
-                        )
-                            .then((onChainAddress: string) => {
-                                runInAction(() => {
-                                    this.onChainAddress = onChainAddress;
-                                    this.payment_request = paymentRequest;
-                                    this.loading = false;
-                                    this.creatingInvoice = false;
-                                });
-                                return { rHash, onChainAddress };
-                            })
-                            .catch(() => {
-                                runInAction(() => {
-                                    this.loading = false;
-                                    this.creatingInvoice = false;
-                                });
-                            });
-                    } else {
-                        runInAction(() => {
-                            this.payment_request = paymentRequest;
-                            this.creatingInvoice = false;
-                        });
-                        return { rHash };
-                    }
-                }
-            )
-            .catch((error: any) => {
+        }).then((result) => {
+            if (!result?.rHash || !result?.paymentRequest) {
                 runInAction(() => {
                     this.loading = false;
                     this.creatingInvoice = false;
                 });
-                throw error;
-            });
+                return Promise.reject(
+                    new Error(
+                        localeString(
+                            'stores.InvoicesStore.errorCreatingInvoice'
+                        )
+                    )
+                );
+            }
+            const { rHash, paymentRequest } = result;
+            if (BackendUtils.supportsOnchainReceiving() && !skipOnchain) {
+                return this.getNewAddress(
+                    addressType
+                        ? { type: addressType, unified: true }
+                        : { unified: true }
+                )
+                    .then((onChainAddress: string) => {
+                        runInAction(() => {
+                            this.onChainAddress = onChainAddress;
+                            this.payment_request = paymentRequest;
+                            this.loading = false;
+                            this.creatingInvoice = false;
+                        });
+                        return { rHash, onChainAddress };
+                    })
+                    .catch(() => {
+                        runInAction(() => {
+                            this.loading = false;
+                            this.creatingInvoice = false;
+                        });
+                    });
+            } else {
+                runInAction(() => {
+                    this.payment_request = paymentRequest;
+                    this.creatingInvoice = false;
+                });
+                return { rHash };
+            }
+        });
     };
 
     @action
@@ -329,7 +318,7 @@ export default class InvoicesStore {
                                         'stores.InvoicesStore.errorCreatingInvoice'
                                     );
                             });
-                            return Promise.reject(error);
+                            return;
                         }
                     }
                     const remotePolicy =
@@ -401,11 +390,7 @@ export default class InvoicesStore {
 
         if (this.lspStore.flow_error) {
             this.creatingInvoice = false;
-            return Promise.reject(
-                new Error(
-                    localeString('stores.InvoicesStore.errorCreatingInvoice')
-                )
-            );
+            return;
         }
 
         return BackendUtils.createInvoice(req)
@@ -429,7 +414,6 @@ export default class InvoicesStore {
                                       'stores.InvoicesStore.errorCreatingInvoice'
                                   );
                     });
-                    return;
                 }
                 const invoice = new Invoice(data);
 
@@ -514,7 +498,6 @@ export default class InvoicesStore {
                             'stores.InvoicesStore.errorCreatingInvoice'
                         );
                 });
-                throw error;
             });
     };
