+++ stores/LSPStore.ts
@@ -427,8 +427,8 @@ export default class LSPStore {
     };
 
     @action
-    public handleCustomMessages = (decoded: any) => {
-        if (!decoded?.peer || !decoded?.data) return;
+    public handleCustomMessages = (decoded: any): boolean => {
+        if (!decoded?.peer || !decoded?.data) return false;
         let data: any;
         try {
             data = JSON.parse(Base64Utils.base64ToUtf8(decoded.data));
@@ -437,7 +437,7 @@ export default class LSPStore {
                 'Custom message data is not valid JSON, skipping:',
                 parseError?.message
             );
-            return;
+            return false;
         }
         const peer = Base64Utils.base64ToHex(decoded.peer);
 
@@ -446,6 +446,7 @@ export default class LSPStore {
         if (data.id === this.getInfoId) {
             this.getInfoData = data;
             this.loadingLSPS1 = false;
+            return true;
         } else if (data.id === this.createOrderId) {
             if (data.error) {
                 this.error = true;
@@ -456,6 +457,7 @@ export default class LSPStore {
                 this.createOrderResponse = data;
             }
             this.loadingLSPS1 = false;
+            return true;
         } else if (data.id === this.getOrderId) {
             if (data.error) {
                 this.error = true;
@@ -466,6 +468,7 @@ export default class LSPStore {
                 this.getOrderResponse = data;
             }
             this.loadingLSPS1 = false;
+            return true;
         } else if (data.id === this.getExtendableOrdersId) {
             if (data.error) {
                 this.error = true;
@@ -476,6 +479,7 @@ export default class LSPStore {
                 this.getExtendableOrdersData = data?.result?.extendable_orders;
             }
             this.loadingLSPS7 = false;
+            return true;
         } else if (data.id === this.createExtensionOrderId) {
             if (data.error) {
                 this.error = true;
@@ -486,6 +490,7 @@ export default class LSPStore {
                 this.createExtensionOrderResponse = data;
             }
             this.loadingLSPS7 = false;
+            return true;
         } else if (data.id === this.getExtensionOrderId) {
             if (data.error) {
                 this.error = true;
@@ -496,7 +501,9 @@ export default class LSPStore {
                 this.getExtensionOrderResponse = data;
             }
             this.loadingLSPS7 = false;
+            return true;
         }
+        return false;
     };
 
     @action
@@ -522,11 +529,13 @@ export default class LSPStore {
                     if (!event?.data) return;
                     try {
                         const decoded = index.decodeCustomMessage(event.data);
-                        runInAction(() => {
-                            this.handleCustomMessages(decoded);
+                        const handled = runInAction(() =>
+                            this.handleCustomMessages(decoded)
+                        );
+                        if (handled) {
                             this.resolvedCustomMessage = true;
-                        });
-                        clearTimeout(timeoutId);
+                            clearTimeout(timeoutId);
+                        }
                     } catch (error: any) {
                         console.error(
                             'sub custom messages error: ' + error.message
@@ -540,11 +549,13 @@ export default class LSPStore {
             BackendUtils.subscribeCustomMessages(
                 (response: any) => {
                     const decoded = response.result;
-                    runInAction(() => {
-                        this.handleCustomMessages(decoded);
+                    const handled = runInAction(() =>
+                        this.handleCustomMessages(decoded)
+                    );
+                    if (handled) {
                         this.resolvedCustomMessage = true;
-                    });
-                    clearTimeout(timeoutId);
+                        clearTimeout(timeoutId);
+                    }
                 },
                 (error: any) => {
                     console.error(
