+++ utils/LndMobileUtils.ts
@@ -36,6 +36,31 @@ import {
 
 import { lnrpc } from '../proto/lightning';
 
+import {
+    LndErrorCode,
+    createLndError,
+    getErrorMessage,
+    isLndError,
+    isStopLndExpectedError,
+    matchesLndErrorCode
+} from './LndMobileErrors';
+
+export {
+    LndErrorCode,
+    LND_ERROR_MESSAGES,
+    LND_ERROR_PATTERNS,
+    MAX_TRANSIENT_RPC_RETRIES,
+    TRANSIENT_RPC_RETRY_BASE_MS,
+    createLndError,
+    getErrorMessage,
+    isLndError,
+    isStopLndExpectedError,
+    isTransientRpcError,
+    matchRawErrorToCode,
+    matchesLndErrorCode,
+    retryOnTransientError
+} from './LndMobileErrors';
+
 // ---------------------------------------------------------------------------
 // Timing constants
 // ---------------------------------------------------------------------------
@@ -45,204 +70,12 @@ const LND_RETRY_DELAY_MS = 3000; // Base delay between LND start retry attempts
 const ANDROID_PROCESS_CLEANUP_DELAY_MS = 4000; // Extra time Android needs for process cleanup (slower than iOS)
 const IOS_PROCESS_CLEANUP_DELAY_MS = 2000; // iOS process cleanup delay
 const GEN_SEED_STOP_DELAY_MS = 3000; // Delay after stopping LND before wallet creation restart attempts
+const GEN_SEED_MAX_RETRIES = 10; // Max retries when LND unlocks too quickly during wallet creation
+const GEN_SEED_RETRY_DELAY_MS = 500; // Delay between genSeed retry attempts (ms)
 const STOP_LND_MAX_RETRIES = 10; // Stop LND: max polling attempts to verify shutdown
 const STOP_LND_POLL_DELAY_MS = 500; // Stop LND: delay between polling attempts (ms)
 const MAX_START_LND_RETRIES = 10; // Maximum start attempts for LND
-
-/**
- * LND error codes - use for consistent error handling across the app
- */
-export enum LndErrorCode {
-    /** LND data folder does not exist */
-    LND_FOLDER_MISSING = 'LND_FOLDER_MISSING',
-    /** LND unlocked too quickly during wallet creation */
-    WALLET_CREATION_UNLOCKED_TOO_QUICKLY = 'WALLET_CREATION_UNLOCKED_TOO_QUICKLY',
-    /** RPC connection was closed */
-    RPC_CONNECTION_CLOSED = 'RPC_CONNECTION_CLOSED',
-    /** RPC did not become ready within timeout */
-    RPC_READY_TIMEOUT = 'RPC_READY_TIMEOUT',
-    /** LND failed to stop after max retries */
-    LND_FAILED_TO_STOP = 'LND_FAILED_TO_STOP',
-    /** LND failed to start after max retries */
-    LND_START_FAILED = 'LND_START_FAILED',
-    /** Failed to generate seed */
-    GEN_SEED_FAILED = 'GEN_SEED_FAILED',
-    /** Timeout waiting for LND to become ready */
-    LND_READY_TIMEOUT = 'LND_READY_TIMEOUT',
-    /** LND is already running (informational) */
-    LND_ALREADY_RUNNING = 'LND_ALREADY_RUNNING',
-    /** Wallet is locked - unlock required */
-    WALLET_LOCKED = 'WALLET_LOCKED',
-    /** Macaroon store is locked */
-    MACAROON_STORE_LOCKED = 'MACAROON_STORE_LOCKED',
-    /** Wallet/Unlocker already unlocked (genSeed race) */
-    GEN_SEED_UNLOCKED = 'GEN_SEED_UNLOCKED',
-    /** RPC not ready yet (transient) */
-    RPC_NOT_READY = 'RPC_NOT_READY',
-    /** Expected state when stopping LND  */
-    STOP_LND_EXPECTED = 'STOP_LND_EXPECTED',
-    /** Invalid response from LND stream */
-    STREAM_INVALID_RESPONSE = 'STREAM_INVALID_RESPONSE',
-    /** Stream ended normally (EOF) - not an error */
-    STREAM_EOF = 'STREAM_EOF'
-}
-
-/**  messages for display in UI */
-export const LND_ERROR_MESSAGES: Record<LndErrorCode, string> = {
-    [LndErrorCode.LND_FOLDER_MISSING]:
-        'LND wallet folder not found. The wallet may have been deleted.',
-    [LndErrorCode.WALLET_CREATION_UNLOCKED_TOO_QUICKLY]:
-        'Wallet creation failed: LND unlocked too quickly. Please try again.',
-    [LndErrorCode.RPC_CONNECTION_CLOSED]: 'RPC connection closed.',
-    [LndErrorCode.RPC_READY_TIMEOUT]:
-        'RPC did not become ready in time. Please try again.',
-    [LndErrorCode.LND_FAILED_TO_STOP]:
-        'LND failed to stop. Please restart the app.',
-    [LndErrorCode.LND_START_FAILED]:
-        'LND failed to start. Please restart the app.',
-    [LndErrorCode.GEN_SEED_FAILED]: 'Failed to generate wallet seed.',
-    [LndErrorCode.LND_READY_TIMEOUT]:
-        'LND did not become ready in time. Please try again.',
-    [LndErrorCode.LND_ALREADY_RUNNING]: 'LND is already running.',
-    [LndErrorCode.WALLET_LOCKED]: 'Wallet is locked. Unlock required.',
-    [LndErrorCode.MACAROON_STORE_LOCKED]:
-        'Macaroon store is locked. Please try again.',
-    [LndErrorCode.GEN_SEED_UNLOCKED]:
-        'Wallet already unlocked. Please try wallet creation again.',
-    [LndErrorCode.RPC_NOT_READY]: 'RPC not ready yet. Please wait.',
-    [LndErrorCode.STOP_LND_EXPECTED]: 'LND stopped (expected state).',
-    [LndErrorCode.STREAM_INVALID_RESPONSE]:
-        'Invalid response from LND. Please try again.',
-    [LndErrorCode.STREAM_EOF]: 'Stream ended (expected).'
-};
-
-/** Raw message patterns that map to each error code (for matching native/stream errors) */
-export const LND_ERROR_PATTERNS: Record<LndErrorCode, readonly string[]> = {
-    [LndErrorCode.LND_FOLDER_MISSING]: [
-        "doesn't exist",
-        'does not exist',
-        'no such file or directory',
-        'wallet directory not found',
-        'wallet may have been deleted'
-    ],
-    [LndErrorCode.WALLET_CREATION_UNLOCKED_TOO_QUICKLY]: [],
-    [LndErrorCode.RPC_CONNECTION_CLOSED]: ['closed', 'connection closed'],
-    [LndErrorCode.RPC_READY_TIMEOUT]: [],
-    [LndErrorCode.LND_FAILED_TO_STOP]: [],
-    [LndErrorCode.LND_START_FAILED]: [],
-    [LndErrorCode.GEN_SEED_FAILED]: [],
-    [LndErrorCode.LND_READY_TIMEOUT]: [],
-    [LndErrorCode.LND_ALREADY_RUNNING]: ['already started', 'already running'],
-    [LndErrorCode.WALLET_LOCKED]: ['wallet locked'],
-    [LndErrorCode.MACAROON_STORE_LOCKED]: [
-        'macaroon store is locked',
-        'cannot retrieve macaroon',
-        'cannot get macaroon'
-    ],
-    [LndErrorCode.GEN_SEED_UNLOCKED]: [
-        'wallet already unlocked',
-        'WalletUnlocker service is no longer available'
-    ],
-    [LndErrorCode.RPC_NOT_READY]: [
-        'starting up',
-        'not yet ready to accept calls',
-        'not yet ready',
-        'unitialized',
-        'uninitialized'
-    ],
-    [LndErrorCode.STOP_LND_EXPECTED]: [
-        'unable to read TLS cert',
-        'connection refused',
-        'connection reset'
-    ],
-    [LndErrorCode.STREAM_INVALID_RESPONSE]: [],
-    [LndErrorCode.STREAM_EOF]: [
-        'EOF',
-        'error reading from server: EOF',
-        'channel event store shutting down'
-    ]
-};
-
-/** Match a raw error message to an LndErrorCode (returns first match) */
-export function matchRawErrorToCode(msg: string): LndErrorCode | null {
-    const normalized = msg.toLowerCase();
-    for (const [code, patterns] of Object.entries(LND_ERROR_PATTERNS)) {
-        if (
-            patterns.length > 0 &&
-            patterns.some((p) => normalized.includes(p.toLowerCase()))
-        ) {
-            return code as LndErrorCode;
-        }
-    }
-    return null;
-}
-
-/** Check if message matches a specific error code */
-export function matchesLndErrorCode(msg: string, code: LndErrorCode): boolean {
-    const patterns = LND_ERROR_PATTERNS[code];
-    if (!patterns.length) return false;
-    const normalized = msg.toLowerCase();
-    return patterns.some((p) => normalized.includes(p.toLowerCase()));
-}
-
-/** Codes that indicate expected state when stopping LND  */
-const STOP_LND_EXPECTED_CODES: LndErrorCode[] = [
-    LndErrorCode.STOP_LND_EXPECTED,
-    LndErrorCode.WALLET_LOCKED,
-    LndErrorCode.RPC_CONNECTION_CLOSED,
-    LndErrorCode.LND_FOLDER_MISSING,
-    LndErrorCode.RPC_NOT_READY
-];
-
-/** Check if error is expected when stopping LND  */
-export function isStopLndExpectedError(msg: string): boolean {
-    return STOP_LND_EXPECTED_CODES.some((code) =>
-        matchesLndErrorCode(msg, code)
-    );
-}
-
-/** Transient RPC errors that may resolve with retry (node switch, startup, etc.) */
-const TRANSIENT_RPC_ERROR_CODES: LndErrorCode[] = [
-    LndErrorCode.RPC_CONNECTION_CLOSED,
-    LndErrorCode.RPC_NOT_READY,
-    LndErrorCode.MACAROON_STORE_LOCKED
-];
-
-/** Check if error message is a transient RPC error (retryable) */
-export function isTransientRpcError(msg: string): boolean {
-    return TRANSIENT_RPC_ERROR_CODES.some((code) =>
-        matchesLndErrorCode(msg, code)
-    );
-}
-
-export function getLndErrorMessage(error: unknown): string {
-    if (isLndError(error)) {
-        const code = (error as { code?: LndErrorCode })?.code;
-        if (code && LND_ERROR_MESSAGES[code]) {
-            return LND_ERROR_MESSAGES[code];
-        }
-    }
-    return error instanceof Error ? error.message : String(error ?? '');
-}
-
-export function createLndError(code: LndErrorCode, detail?: string): Error {
-    const baseMessage = LND_ERROR_MESSAGES[code];
-    const message = detail ? `${baseMessage} ${detail}`.trim() : baseMessage;
-    const err = new Error(message) as Error & { code: LndErrorCode };
-    err.code = code;
-    return err;
-}
-
-export function isLndError(error: unknown, code?: LndErrorCode): boolean {
-    const msg = error instanceof Error ? error.message : String(error ?? '');
-    const errCode = (error as { code?: LndErrorCode })?.code;
-    if (code) {
-        return errCode === code || msg.includes(code);
-    }
-    return Object.values(LndErrorCode).some((c) => msg.includes(c));
-}
-
-/** Android needs longer delays - process cleanup is slower */
+const LND_READY_TIMEOUT_MS = 60000; // Max wait for LND to reach ready state (wallet/RPC)
 
 export const NEUTRINO_PING_TIMEOUT_MS = 1500;
 export const NEUTRINO_PING_OPTIMAL_MS = 200;
@@ -494,15 +327,6 @@ export async function initializeLnd({
     await initialize();
 }
 
-function getErrorMessage(error: unknown): string {
-    if (error instanceof Error) return error.message;
-    if (typeof error === 'string') return error;
-    if (error && typeof error === 'object' && 'message' in error) {
-        return String((error as { message: unknown }).message);
-    }
-    return String(error);
-}
-
 /**
  * Stops the LND process gracefully with retry mechanism
  * @param maxRetries - Maximum number of polling attempts to verify shutdown (default: 10)
@@ -580,7 +404,6 @@ export async function stopLnd(
         // Poll until LND process has fully stopped
         await retry({
             fn: async () => {
-                await sleep(delayMs);
                 const currentStatus = await runWithExpectedErrorHandling(
                     () => checkStatus(),
                     'poll'
@@ -602,7 +425,7 @@ export async function stopLnd(
                 );
             },
             maxRetries,
-            delayMs: 0,
+            delayMs,
             onRetry: (attempt) => {
                 log.d(
                     `Stop polling attempt ${attempt}/${maxRetries}, LND still running`
@@ -825,6 +648,7 @@ async function retryStartLnd({
 /**
  * Helper function to wait for LND to reach a ready state
  */
+
 async function waitForLndReady({
     decodeState,
     subscribeState,
@@ -835,31 +659,51 @@ async function waitForLndReady({
     subscribeState: () => Promise<string>;
     walletPassword: string;
     unlockWallet: (password: string) => Promise<void>;
-}) {
-    return new Promise(async (resolve, reject) => {
-        let isResolved = false;
-        let unlockAttempted = false;
+}): Promise<boolean> {
+    let unlockAttempted = false;
 
-        const stateHandler = async (event: any) => {
-            if (isResolved) return;
+    const handleRpcReady = async (context: string) => {
+        await waitForRpcReady();
+        if (walletPassword && !syncStore.isSyncing) {
+            log.d('Starting sync');
+            syncStore.startSyncing();
+        }
+        if (settingsStore?.settings?.rescan) {
+            syncStore.startRescanTracking(0);
+        }
+        log.d(`${context}: RPC ready`);
+    };
 
-            try {
+    const waitForState = () =>
+        new Promise<boolean>((resolve, reject) => {
+            let settled = false;
+
+            const cleanup = () => {
+                clearTimeout(timeout);
+                LndMobileEventEmitter.removeAllListeners('SubscribeState');
+            };
+
+            const settle = (fn: () => void) => {
+                if (settled) return;
+                settled = true;
+                cleanup();
+                fn();
+            };
+
+            const stateHandler = (event: any) => {
+                if (settled) return;
                 log.d('SubscribeState event received', [event]);
 
                 const error = checkLndStreamErrorResponse(
                     'SubscribeState',
                     event
                 );
-                if (error === 'EOF') {
-                    return;
-                } else if (error) {
-                    isResolved = true;
-                    cleanup();
-                    reject(error);
+                if (error === 'EOF') return;
+                if (error) {
+                    settle(() => reject(error));
                     return;
                 }
 
-                // Skip decode when connection closed or no data (avoids uncaught errors)
                 if (event?.error_code && !event?.data) {
                     log.d(
                         'SubscribeState: connection closed or no data, skipping'
@@ -870,127 +714,78 @@ async function waitForLndReady({
                 const state = decodeState(event.data ?? '');
                 log.d(`Current LND state: ${state.state}`);
 
-                switch (state.state) {
-                    case lnrpc.WalletState.NON_EXISTING:
-                        log.d('Wallet does not exist - ready for creation');
-                        isResolved = true;
-                        cleanup();
-                        await sleep(STATE_SUBSCRIPTION_SETTLE_MS);
-                        resolve(true);
-                        break;
+                const handleAsync = async () => {
+                    switch (state.state) {
+                        case lnrpc.WalletState.NON_EXISTING:
+                            log.d('Wallet does not exist - ready for creation');
+                            await sleep(STATE_SUBSCRIPTION_SETTLE_MS);
+                            settle(() => resolve(true));
+                            break;
 
-                    case lnrpc.WalletState.LOCKED:
-                        log.d('Wallet is locked');
-                        if (!unlockAttempted && walletPassword) {
-                            unlockAttempted = true;
-                            try {
+                        case lnrpc.WalletState.LOCKED:
+                            log.d('Wallet is locked');
+                            if (!unlockAttempted && walletPassword) {
+                                unlockAttempted = true;
                                 await unlockWallet(walletPassword);
                                 log.d('Wallet unlocked successfully');
-                            } catch (unlockError: any) {
-                                log.e('Error unlocking wallet', [unlockError]);
-                                isResolved = true;
-                                cleanup();
-                                reject(unlockError);
-                                return;
                             }
-                        }
-                        break;
+                            break;
 
-                    case lnrpc.WalletState.UNLOCKED:
-                        log.d(
-                            'Wallet unlocked - waiting for RPC to become active'
-                        );
-                        // Don't resolve - wait for RPC_ACTIVE or SERVER_ACTIVE
-                        break;
+                        case lnrpc.WalletState.UNLOCKED:
+                            log.d(
+                                'Wallet unlocked - waiting for RPC to become active'
+                            );
+                            break;
 
-                    case lnrpc.WalletState.RPC_ACTIVE:
-                        log.d('RPC is active - waiting for RPC ready');
-                        try {
-                            await waitForRpcReady();
-                            // Only start sync when unlocking existing wallet - not during wallet creation
-                            if (walletPassword && !syncStore.isSyncing) {
-                                syncStore.startSyncing();
+                        case lnrpc.WalletState.RPC_ACTIVE:
+                            log.d('RPC is active');
+                            await handleRpcReady('RPC_ACTIVE');
+                            settle(() => resolve(true));
+                            break;
+
+                        case lnrpc.WalletState.SERVER_ACTIVE:
+                            log.d('Server is active');
+                            try {
+                                await handleRpcReady('SERVER_ACTIVE');
+                            } catch (e: any) {
+                                log.e(
+                                    'RPC ready check failed (SERVER_ACTIVE)',
+                                    [e]
+                                );
                             }
-                            if (settingsStore?.settings?.rescan) {
-                                syncStore.startRescanTracking(0);
-                            }
-                            isResolved = true;
-                            cleanup();
-                            resolve(true);
-                        } catch (rpcError: any) {
-                            log.e('RPC ready check failed', [rpcError]);
-                            isResolved = true;
-                            cleanup();
-                            reject(rpcError);
-                            return;
-                        }
-                        break;
+                            settle(() => resolve(true));
+                            break;
 
-                    case lnrpc.WalletState.SERVER_ACTIVE:
-                        log.d('Server is active');
-                        try {
-                            await waitForRpcReady();
-                            if (walletPassword && !syncStore.isSyncing) {
-                                log.d('Starting sync');
-                                syncStore.startSyncing();
-                            }
-                            if (settingsStore?.settings?.rescan) {
-                                syncStore.startRescanTracking(0);
-                            }
-                        } catch (rpcError: any) {
-                            log.e('RPC ready check failed (SERVER_ACTIVE)', [
-                                rpcError
-                            ]);
-                        }
-                        isResolved = true;
-                        cleanup();
-                        resolve(true);
-                        break;
+                        default:
+                            log.d('Unknown wallet state', [state.state]);
+                            break;
+                    }
+                };
 
-                    default:
-                        log.d('Unknown wallet state', [state.state]);
-                        break;
-                }
-            } catch (error: any) {
-                log.e('SubscribeState handler error', [error]);
-                if (!isResolved) {
-                    isResolved = true;
-                    cleanup();
-                    reject(error);
-                }
-            }
-        };
+                handleAsync().catch((error) => {
+                    log.e('SubscribeState handler error', [error]);
+                    settle(() => reject(error));
+                });
+            };
 
-        const cleanup = () => {
-            if (timeout) clearTimeout(timeout);
-            LndMobileEventEmitter.removeAllListeners('SubscribeState');
-        };
+            const timeout = setTimeout(
+                () =>
+                    settle(() =>
+                        reject(createLndError(LndErrorCode.LND_READY_TIMEOUT))
+                    ),
+                LND_READY_TIMEOUT_MS
+            );
 
-        const timeout = setTimeout(() => {
-            if (!isResolved) {
-                isResolved = true;
-                cleanup();
-                reject(createLndError(LndErrorCode.LND_READY_TIMEOUT));
-            }
-        }, 60000);
+            LndMobileEventEmitter.addListener('SubscribeState', stateHandler);
+        });
 
-        LndMobileEventEmitter.addListener('SubscribeState', stateHandler);
+    await sleep(LISTENER_REGISTRATION_MS);
 
-        // Give the listener time to fully register
-        await sleep(LISTENER_REGISTRATION_MS);
+    log.d('Starting state subscription');
+    await subscribeState();
+    log.d('State subscription started successfully');
 
-        try {
-            log.d('Starting state subscription');
-            await subscribeState();
-            log.d('State subscription started successfully');
-        } catch (error) {
-            if (!isResolved) {
-                isResolved = true;
-                cleanup();
-                reject(error);
-            }
-        }
-    });
+    return waitForState();
 }
 
 export async function optimizeNeutrinoPeers(
@@ -1002,7 +797,7 @@ export async function optimizeNeutrinoPeers(
         ? DEFAULT_NEUTRINO_PEERS_TESTNET
         : DEFAULT_NEUTRINO_PEERS_MAINNET;
 
-    const results: any = [];
+    const results: { peer: string; ms: number | string }[] = [];
     for (let i = 0; i < peers.length; i++) {
         const peer = peers[i];
         await new Promise(async (resolve) => {
@@ -1185,47 +980,56 @@ async function restartLndForWalletCreation(
 ): Promise<void> {
     const { startLnd, decodeState, subscribeState } = lndMobile.index;
     const { unlockWallet } = lndMobile.wallet;
-    const maxAttempts = 10;
 
-    for (let i = 0; i < maxAttempts; i++) {
-        log.d(
-            `Restarting LND for wallet creation (attempt ${
-                i + 1
-            }/${maxAttempts})`
-        );
-        await stopLnd();
-        const delay = GEN_SEED_STOP_DELAY_MS + i * 1000;
-        await sleep(delay);
+    await stopLnd();
+    await sleep(GEN_SEED_STOP_DELAY_MS);
 
-        try {
-            await startLnd({
-                args: '',
-                lndDir,
-                isTorEnabled: false,
-                isTestnet
-            });
-            await sleep(500);
-            await waitForLndReady({
-                decodeState,
-                subscribeState,
-                walletPassword: '',
-                unlockWallet
-            });
-            return;
-        } catch (err: any) {
-            const msg = getErrorMessage(err);
-            if (matchesLndErrorCode(msg, LndErrorCode.LND_ALREADY_RUNNING)) {
+    try {
+        await retry({
+            fn: async () => {
+                await startLnd({
+                    args: '',
+                    lndDir,
+                    isTorEnabled: false,
+                    isTestnet
+                });
+                await waitForLndReady({
+                    decodeState,
+                    subscribeState,
+                    walletPassword: '',
+                    unlockWallet
+                });
+            },
+            maxRetries: MAX_START_LND_RETRIES,
+            delayMs: 0,
+            shouldRetry: (e) =>
+                matchesLndErrorCode(
+                    getErrorMessage(e),
+                    LndErrorCode.LND_ALREADY_RUNNING
+                ),
+            onRetry: async (attempt) => {
                 log.d(
-                    `LND still running after stop, retrying with longer delay (${
-                        GEN_SEED_STOP_DELAY_MS + (i + 1) * 1000
-                    }ms)`
+                    `Restarting LND for wallet creation (attempt ${
+                        attempt + 1
+                    }/${MAX_START_LND_RETRIES})`
                 );
-                continue;
+                await stopLnd();
+                await sleep(GEN_SEED_STOP_DELAY_MS + attempt * 1000);
             }
-            throw err;
+        });
+    } catch (e) {
+        if (
+            matchesLndErrorCode(
+                getErrorMessage(e),
+                LndErrorCode.LND_ALREADY_RUNNING
+            )
+        ) {
+            throw createLndError(
+                LndErrorCode.WALLET_CREATION_UNLOCKED_TOO_QUICKLY
+            );
         }
+        throw e;
     }
-    throw createLndError(LndErrorCode.WALLET_CREATION_UNLOCKED_TOO_QUICKLY);
 }
 
 /**
@@ -1244,8 +1048,8 @@ async function genSeedWithRetry(
             }
             return seed;
         },
-        maxRetries: 10,
-        delayMs: 500,
+        maxRetries: GEN_SEED_MAX_RETRIES,
+        delayMs: GEN_SEED_RETRY_DELAY_MS,
         shouldRetry: (e) =>
             matchesLndErrorCode(
                 getErrorMessage(e),
